// packages/contracts/generated/NoteView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { IpcMarkerKind } from "./IpcMarkerKind";
import type { IpcNodeKind } from "./IpcNodeKind";

export type NoteView = { id: string, parentId: string | null, sortKey: number, kind: IpcNodeKind, text: string, note: string, marker: IpcMarkerKind, collapsed: boolean, completed: boolean, starred: boolean, deleted: boolean, };